QSDK:CP: Fix issue in updating the VHT, HE and HT IE's
When the chanmask values are being updated using the following

$ iw phy set antenna TX RX

there is change  in the values stored in ar->cfg_tx_chainmask and
ar->cfg_rx_chainmask. These values are used for updating the IE's
and also setting the respective enabled bit. Since the value
stored can be any user configured value not only the max supported by
the hardware we will have issue when user configures values like 4,5,6, etc.
i.e non continue 11's and so the IE's will be updated incorrectly

E.g lets say user configures the values as 4 for tx and rx
i.e ar->num_tx_chains=1 and rate_cap_tx_chainmask= 0100

      if (i < ar->num_tx_chains && rate_cap_tx_chainmask & BIT(i))
		txmcs_map |= IEEE80211_VHT_MCS_SUPPORT_0_9 << (i * 2);
      else
		txmcs_map |= IEEE80211_VHT_MCS_NOT_SUPPORTED << (i * 2);

now based on the loop we will not set the VHT_MCS_SUPPORTED_0_9 for any
of the mask so changing the loop as below
      if (i < ar->num_tx_chains)
		txmcs_map |= IEEE80211_VHT_MCS_SUPPORT_0_9 << (i * 2);
      else
		txmcs_map |= IEEE80211_VHT_MCS_NOT_SUPPORTED << (i * 2);

Above loop will  enable 1SS for VHT_MCS_SUPPORTED_0_9 and rest disabled.
Similar behaviour is seen in other places also so correcting it across
the code flow

Change-Id: Ideb483d2db4be7a425e4ab4325224e282cdcdd3d
Signed-off-by: Uraj Sasan <usasan@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>