UPSTREAM:DP: wifi: ath12k: Add DP UMAC Reset support
It contains the base changes to add support for umac dp
reset of partner SOC when one SOC crashes in MLO
as a part of MODE 1 SSR

In this process shared memory will be used by Host and
Firmware to communicate the sequence of events. Host will
clear the data structures such as tx/rx descriptors and
re-initialize the rings and reset hp/tp to 0.

Following patches are partially implemented as part of this change:
(Partially implemented as few of the patches are already
added as part of package upgrade)

800-wifi-ath12k-Add-UMAC-DP-reset-support.patch (Partial)
850-03-wifi-ath12k-avoid-pass-the-number-of-entry-to-replen.patch
850-06-wifi-ath12k-avoid-used-list.patch (partial)
955-02-ath12k-simplify-rx-buffer-ring-replenish-logic.patch (Partial)

Change-Id: Ibaeb2e1e331c0a4e9457cee44a8a6b1cf92c2322
Signed-off-by: Ramya Gnanasekar <rgnanase@qti.qualcomm.com>
Signed-off-by: Karthikeyan Periyasamy <periyasa@qti.qualcomm.com>
Signed-off-by: P Praneesh <ppranees@qti.qualcomm.com>
Signed-off-by: Adwait Nayak <adwaitn@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>