UPSTREAM:PROTOCOL:wifi:ath12k: Update wiphy sband for 2GHz if duplicated
When 2GHz band capabilities VHT_EXT_NSS_BW_CAPABLE is checked and if the support
is not there wiphy duplicates the sband info in ieee80211_register_hw().
Due to this the wiphy sband between the mac layer and driver is not same.

Now when the user changes the tx/rx chinmask the VHT_MCS is updated but
since the sband info stored in the driver is different from the sband info
in the wiphy the userspace does not receive the updated vht_mcs values.

Make sure to update vhtmcs when sband is duplicated during chainmask change.

Change-Id: I00281378174124031e522a2a6d08675d663b9c45
Signed-off-by: Uraj Sasan <usasan@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>