UPSTREAM:PROTOCOL:wifi:ath11k:Add support for Tx/Rx chainmask based on radio id
Add support to add the radio id parameter for which Tx/Rx chainmask
needs to be changed from userspace. Pass on this radio index to
underlying drivers as an additional argument.

A value of NL80211_WIPHY_RADIO_ID_MAX(255) indicates radio index is not
mentioned and the configuration applies to all radio(s) of the wiphy.

Change-Id: I1b1ba4cfbc18997f633c22bc73462d22f0724f9c
Signed-off-by: Uraj Sasan <usasan@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>