UPSTREAM: DP: wifi: ath12k: Drop packets with invalid rate info in Rx path
The following issues are addressed in this patch:

1. Rarely we see frames received with invalid channel number
in rx descriptor. Fetching the band based on this channel
leads to warning in mac80211. In such cases dump the rx desc
to identify the channel/freq received from HW.

2. If sband is NULL during ath12k_wifi7_dp_rx_h_ppdu call for 6G band, then
rx_status->freq is not filled. Due to this, facing panic_on_warn_on
in ieee80211_rx_monitor function when trying to get hw_idx from the
same frequency value. Fixing this by properly filling the
rx_status->freq value in sband NULL case.

3. Currently, some packets received with invalid MCS  trigger a WARN_ON.
This is because some clients support higher capabilities than the AP,
allowing them to send packets in higher MCS 12 and 13 rates, causing
this issue. Hence, add a WARN_ON_ONCE in driver and drop the invalid
mcs packets.

4. Currently, warn_on is getting trigger due to the presence of
corrupted packets associated with an invalid band from the
WBM error ring.
This issue can lead to unnecessary warnings and potential disruptions

Hence, add a check to drop the packets if it has invalid band and
originating from the WBM error path.

This will ensure corrupted packets are dropped and prevent the
warn_on trigger.

5. Move check for rx_desc_data->is_drop_packet after API
ath12k_wifi7_dp_rx_h_mpdu() in regular Rx processing as flag is set
inside this API.

Change-Id: I72add3d19c71374e4b38a80a3d9cdb89a1d67877
Signed-off-by: Sriram R <srirrama@qti.qualcomm.com>
Signed-off-by: Sarika Sharma <sarishar@qti.qualcomm.com>
Signed-off-by: Thiraviyam Mariyappan <tmariyap@qti.qualcomm.com>
Signed-off-by: Vinith Kumar R <vinithku@qti.qualcomm.com>
Signed-off-by: Harsh Kumar Bijlani <hbijlani@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>