QSDK:CP: Adevertise recovery Mode as Mode-0/Mode-1 to FW
By default FW will operate in Mode-0 recovery Mode which means
whenever there is a crash in Q6 the partner device crash indication
is sent to all the Q6 that are part of the MLO. But for Mode-1
recovery only the asserted readio should go for a Q6 restart.
To achieve this and to toggle between Mode-0 and Mode-1 recovery
modes FW rely on MLO global memory allocated during QMI seqeunce.

This will be used a Shared mem between FW and host to advertise
the recovery mode as Mode-0/Mode-1. Added changes to parse this MLO
memory and get the address of shared memory and poulate in ath12k_base.

Also add support to send the crash notification as userpd crash
for AHB/HYBRID chipsets.

Add support to trigger umac reset in case of Mode-1 recovery
and also for the cleanup of ML peer in non asserted partner
chipsets.

Change-Id: I2afb41b0d44b06d89f0e66c1f8e360cff0218043
Signed-off-by: Manish Dharanenthiran <mdharane@qti.qualcomm.com>
Signed-off-by: Vignesh C <vignc@qti.qualcomm.com>
Signed-off-by: Sidhanta Sahu <sidhanta@qti.qualcomm.com>
Signed-off-by: Sowmiya Sree Elavalagan <ssreeela@qti.qualcomm.com>
Signed-off-by: Avula Sri Charan <asrichar@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>