UPSTREAM:PROTOCOL:cfg80211: Enhance 6 GHz Channel List Handling
Currently, the wiphy sbands for 6 GHz band has 2 set of channels
  (1) Channels array for the 6 GHz band which is power mode agnostic
  (2) Power mode channel list for each of the 9 different power modes.

Once the driver sends an indication to the cfg80211 with a new regd,
for the 6 GHz band, the power mode channel list is modified based on
the new regd information for each power mode.

However, the power mode agnostic channel list is not updated
accordingly, which can lead to inconsistencies.

Add logic to fill the power mode agnostic channel list based on the
power mode channel list. A channel will be treated as enabled in the
power mode agnostic channel list if it is enabled in any of the power
mode channel lists.

This ensures that the power mode agnostic channel list is always
a superset of the power mode channel lists. This is important for
legacy callers of the cfg80211 API who are not aware of the power mode
and are only concerned about the enablement of the 6 GHz channel
irrespective of the power mode.

Fixes: None
Patch-work: None
Tag: 6G_REGULATORY
Feature: AFC
Patch-dependency:
subsys/0048-cfg80211-mac80211-regulatory-awgn-and-6g-channel-cha.patch

Change-Id: I460dc735ba6374574834230507ccdbcb1001625d
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>