UPSTREAM:PROTOCOL:wifi:ath12k: 6 GHz power mode change sanity check
Currently, there is no sanity check for the 6 GHz power mode change
command to prevent it from getting set on an interface which is
down. This causes a NULL pointer dereference in the driver, when
trying to reach the link chandef's channel, which is NULL as the
interface is down.
To fix this, ensure that the 6 GHz power mode change command
is allowed only when the interface is up.

Changing the power mode of the device is only applicable for AP
type interfaces. It does not make sense for other interface types.
Make explicit check to block the power mode change command for
non-AP type interfaces.

Remove the duplicate power mode channel validity check in the
driver as the cfg80211 API 'cfg80211_update_chandef_6ghz_power_mode'
anyway does the same check.

Fixes: None
Patch-work: None
Tag: 6G_REGULATORY
Feature: AFC
Patch-dependency:
subsys/0048-cfg80211-mac80211-regulatory-awgn-and-6g-channel-cha.patch
subsys/0302-01-UPSTREAM-PROTOCOL-cfg80211-Add-Support-for-NO_IR-for.patch
ath12k/0346-02-wifi-ath12k-Add-support-for-processing-AFC-power-eve.patch

Change-Id: I209b98b11b4401ec7e167020d92409c09824a80f
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>