UPSTREAM: DP: wifi: ath12k: Segregation of Monitor pdev Objects in Unified Driver Framework
With the Unified Driver framework, to support different monitor
variants (quad ring and dual ring) model and for better view of
monitor object, segregate monitor-related rings and objects from
the PDEV DP object and defining them in the monitor pdev object.
Allocation and deallocation of the monitor object are done dynamically.

Patch-dependency:
	0402-0003-UPSTREAM-DP-wifi-ath12k-Refactor-ring-allocation-set.patch

Co-developed-by: Alok Singh <aloksing@qti.qualcomm.com>
Change-Id: I2dcac15f430b29ffa15c5deb69c8152a12923e23
Signed-off-by: Alok Singh <aloksing@qti.qualcomm.com>
Signed-off-by: Sushant Butta <sbutta@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>