UPSTREAM: DP: wifi: ath12k: Refactor monitor pdev ring allocation, setup API's
Monitor destination ring is specific to the dual ring architecture
(mon2), the related allocation, initialization, and setup logic resides
in dp_mon.c, which serves as the common file for both WiFi7 and
the upcoming WiFi8 chipset.

The monitor operations (mon ops) are only initialized for mon2,
ensuring that specific functionalities tied to
monitor destination rings are exclusive to the mon2 architecture.
This prevents unnecessary initialization for architectures
that don't require monitor destination rings.

Patch-dependency:
	0402-0004-UPSTREAM-DP-wifi-ath12k-Segregation-of-Monitor-pdev-.patch

Co-developed-by: Alok Singh <aloksing@qti.qualcomm.com>
Change-Id: I4112aa4a8ee989e00d855470095f62a519d61e7b
Signed-off-by: Alok Singh <aloksing@qti.qualcomm.com>
Signed-off-by: Sushant Butta <sbutta@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>