UPSTREAM: DP: Avoid using hal_rx_desc_data in undecap API
Avoid using hal_rx_desc_data in undecap API, to make this
shared API ready for using new hw rx descriptor data structures.

Replicate the Rx HW descriptor to use as rx cb structure in
the upcoming patches. With this, we can copy most of the data
needed in Rx path in a single go, and use them in the rx path
directly instead of copying from tlvs.

Patch-dependency : None

Change-Id: Icc0f7e2889389253362f335e40462ec8a1233d17
Signed-off-by: Pavankumar Nandeshwar <pnandesh@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>