UPSTREAM:PROTOCOL: wifi: cfg80211: Add an ext feature flag to support offload mode of TTLM
Add an extended feature flag
NL80211_EXT_FEATURE_BEACON_ADVERTISED_TTLM_OFFLOAD to advertise the
support to offload advertisement of TTLM in beacons.

Interested driver modules to set this feature flag and handle the TTLM
as part of ieee80211_vif_cfg_change_notify.

This feature flag can be used to offload TTLM handling from mac80211 to the
driver, providing vendors with greater flexibility. The current mac80211
implementation forcibly disables or removes a link if it is not included
in the TTLM value, marking it as dormant or suspended—behavior that may
not align with all vendor requirements.

Ex: When negotiated mapping is heterogeneous (each TID can map to a
different link set), disabling the links in upper layer won't help
in distributing the traffic. There can be a chance that all links
are enabled but are distributed among TIDs in both UL and DL direction.

Change-Id: Ibfde3f78b4e6bc6008545f7bddf037b68fa19a79
Signed-off-by: Shwetha G K <kshwetha@qti.qualcomm.com>
Signed-off-by: Amith A <amitajit@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>