UPSTREAM:PROTOCOL: wifi: ath12k: Handle driver offloaded advertised TTLM request on AP
With the extension of mac80211 to start the offloaded advertised TTLM on AP
interface, the driver gets VIF cfg change request from upper layers. Hence,
enhanced ath12k driver to handle the request and start the advertisemt by
sending via WMI command.

When there is a 6GHz link present in the MLD interface, the command is
preferred to be sent on 6GHz link vdev. If there is no 6GHz link vdev, the
command can be sent on any link of the MLD interface.

Also, set wiphy flag and ext feature flag to advertise the support of
offload of TTLM advertisement in beacons.

Change-Id: Ifd87a0e171fed22446d2df9bb1c3dc574520bb76
Signed-off-by: Shwetha G K <kshwetha@qti.qualcomm.com>
Signed-off-by: Amith A <amitajit@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>