QSDK: DP: wifi: ath12k: Remove usage of skb_cb in Tx completion path
Currently in Tx completion path, we use the skb_cb for paddr, len,
paddr_ext_desc and paddr_ext_len. Accessing this in that path results in
KPI and CPU load.

In order to optimize this, avoid accessing skb_cb in that path and instead
store the same values in SW Tx descriptor ath12k_tx_desc_info and scratch pad
buffer ath12k_tx_sw_metadata and use them in that path.

Change-Id: Ic373c383d23fd420f5674c224ab2baf0b25340d7
Signed-off-by: Ripan Deuri <rdeuri@qti.qualcomm.com>
Signed-off-by: Harsh Kumar Bijlani <hbijlani@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>