UPSTREAM:PROTOCOL:wifi:ath12k: AFC payload cleanup and SP NO_IR changes
Currently, the AFC payload is not freed up during regulatory deinit.
Also, the AFC payload is not cleared when we move to a new country.
This causes leak of memory and also causes issues wherein we
end up using stale AFC payload data when we move to a new country.
To fix this, free the AFC payload during regulatory deinit and
also clear the AFC payload when we move to a new country.

Also, currently, for countries which does not support SP, SP reg
rules are not updated in the kernel regd. This is not right as the
SP reg rules are still valid and are temporarily unavailable for AP
operation. Also, we can allow passive scanning in these channels.
To fix this issue, update the SP reg rules in the kernel regd by
marking the channels as NO_IR.

Fix the data type of the arguments in the declaration of the function
'ath12k_reg_get_opclass_from_bw' to match the data type of the
actual arguments passed to the function.

Currently, the API 'ath12k_reg_intersect_sp_rules()' takes the
regulatory SP rules and AFC SP rules and intersects them to get the
final SP rules. Refactor and simplify this API to copy the flags
and Tx powers from the Regulatory SP rules to the final SP rules
This is done to ensure that the final SP rules store the
correct Tx power and flags for the country irrespective of the
status of AFC completion.

Fixes: None
Patch-work: None
Tag: 6G_REGULATORY
Feature: AFC
Patch-dependency:
subsys/0048-cfg80211-mac80211-regulatory-awgn-and-6g-channel-cha.patch
ath12k/0346-02-wifi-ath12k-Add-support-for-processing-AFC-power-eve.patch
ath12k/0346-06-wifi-ath12k-Extract-AFC-response-and-send-AFC-cmd-to.patch

Change-Id: Id4c61d5d63654504d3a3ff4c2c01327767ec68ad
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>