UPSTREAM:PROTOCOL:wifi:ath12k: Intersect AFC payload for dual 6 GHz radio
Currently, if two 6 GHz radios are present in the same HW group,
the AFC payload received for one radio overwrites the AFC payload
previously received for the other radio. This can lead to incorrect
SP rules being stored in the wiphy regd.

To fix this, intersect the AFC payloads for both radios and combine
the SP rules from both radios into a single set of rules.
This ensures that the final set of SP rules reflects the combined
payloads from both radios, preventing any overwriting issues.

Currently, when a new AFC payload is received in host, the memory for
old AFC payload stored in struct ath12k is cleared and a new memory
is allocated and assigned directly to the ath12k data structure.

This causes concurrent access issues where there is a possibility of
the AFC payload being cleared for a radio, while the same is being
accessed by a partner 6 GHz radio as part of SP rules intersection.

To prevent the concurrent access issues, do not free the AFC payload
stored in struct ath12k when a new AFC payload is received. Also,
allocate a new AFC payload and pass the payload to regulatory
API for validation, upon which the AFC payload is assigned to
the ath12k data structure.

To ensure data consistency and to avoid concurrent access to AFC
payloads of both the radios, take an 'ah' afc lock to serialize the
AFC payload processing in the regulatory handler.
Free the AFC payload for a radio under a radio level lock to
avoid concurrent access to the AFC payload when it is being freed.

Make changes to the prototype of the regulatory handler to
accept the AFC payload as an argument.

Patch-work: None
Tag: 6G_REGULATORY
Feature: AFC
Patch-dependency:
subsys/0048-cfg80211-mac80211-regulatory-awgn-and-6g-channel-cha.patch
ath12k/0346-02-wifi-ath12k-Add-support-for-processing-AFC-power-eve.patch
ath12k/0346-06-wifi-ath12k-Extract-AFC-response-and-send-AFC-cmd-to.patch
ath12k/0380-a-UPSTREAM-PROTOCOL-wifi-ath12k-AFC-payload-cleanup-an.patch
ath12k/0380-b-UPSTREAM-PROTOCOL-wifi-ath12k-Generate-SP-rules-base.patch
ath12k/0380-c-UPSTREAM-PROTOCOL-wifi-ath12k-Generate-NO_IR-rules-b.patch

Change-Id: If5f25afd12d952bcbfa1b0356567244c0d209d17
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>