QSDK: DP: wifi: ath12k: Remove locks over Tx WBM completion rings
Currently, there is spinlock acquired over Tx WBM completion ring in
tx completion path.

This lock is not required since each ring is mapped to 1 core and hence
there is not possibility of race condition among them.

Hence it is safe to remove this lock. Removal of this lock results in
KPI and CPU load improvements.

Change-Id: I808c5a152416bd1e22d8d9e67740b8c17868b928
Signed-off-by: Harsh Kumar Bijlani <hbijlani@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>