QSDK: CP: wifi: ath12k: Add support for SDWF config using vendor commands
Add support for SDWF service class configuration.
Add support in the driver for the following configuration by handling
the different SDWF vendor NL commands
	Service Class Create
	Service Class Disable
	Service Class View.

To create a service class, based on the the DL or UL paramaters, call the
API to configure DL and UL QoS profile and get the DL and UL QoS profile
ID.

The profile ID is then mapped to a service class using the following API
ath12k_sdwf_map_service_class.

Change-Id: I1ae9295fac1b999d95ec0e9d084307d2dae740d6
Signed-off-by: Vivek Chettri <vchettri@qti.qualcomm.com>
Signed-off-by: Aishwarya R <aisr@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>