UPSTREAM: DP: wifi: ath12k: Segregation of Monitor Objects in Unified Driver Framework
With the Unified Driver framework, to support different monitor mode
variants quad ring model(mon1) for wcn family and dual ring model(mon2)
for qcn family and for better view of monitor object, segregating
monitor-related rings and objects from Data path object (DP) and
defining them in the monitor object (dp_mon).
Allocation and deallocation of the monitor object are done dynamically.

This patch (and the forthcoming patches) is intended to serve the
purpose of refactoring DP objects to DP monitor objects
for the Unified Driver framework.

Patch-dependency:
	0401-0013-UPSTREAM-DP-wifi-ath12k-enable-monitor-mode-for-WCN7.patch

Change-Id: I1c464eea686b11e1576fdbf9c829cf7325a00be1
Signed-off-by: Sushant Butta <sbutta@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>