UPSTREAM: CP: wifi: ath12k: Use random mac address for scan vdev
Currently for scanning purpose, there is a separate vdev created and
this vdev uses the same mac address as that of the existing AP vdev.

Due to the same mac address usage, the ath12k_dp_peer of the scan vdev and
AP vdev point to the same ath12k_dp_peer and when this scan vdev is deleted,
then ath12k_dp_peer also gets deleted. This results in use-after-free of
ath12k_dp_peer for AP vdev in the Tx path.

In order to fix this, generate and make use of random mac address for
scan vdev.

Change-Id: I657c4ddefa0345acfe48dda06a3a805b579b833e
Signed-off-by: Harsh Kumar Bijlani <hbijlani@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>