UPSTREAM: DP: wifi: ath12k: Add support for QoS Datapath
Network flows of different SCS request(identified by SCS ID) of a peer
requires different QoS handling based on the QoS characteristics as
exchanged in the protocol for a given SCS request.

With the help of kernel net filter, as programmed using
nftables by hostapd, we get the SCS ID per packet matching the SCS flow.

Active flows from different SCS request would need a mapping
of the QoS profile or TID and its relevant MSDUQ
that is used to achieve the QoS prioritization.

So per peer, we maintain SCS ID to QoS ID and MSDUQ mapping.
Also we maintain information about all active user define MSDUQs
as they used for different SCS flows of a peer.

These mapping are then referenced in the per packet datapath
to enqueue packets from a flow mapped for a SCS request
to the appropriate MSDUQ.

Add API to support the following

	Allocate a peer QoS context to store the active SCS requests
	and its QoS profile, and MSDUQ information

	Add active SCS information of QoS profile and MSDUQ
	in peer QoS context

	Delete terminated SCS information of QoS profile and MSDUQ
	in peer QoS context

	Free peer QoS context when the peer is deleted.

In per packet DP Tx, retrieve the SCS mapping for QoS profile and MSDUQ
and update the TCL command to use the appropriate HW MSDUQ
for transmission for achieving configured QoS.

Change-Id: Ic74b845ceb86c6909a05f4e5140432fccf260ba4
Signed-off-by: Vivek Chettri <vchettri@qti.qualcomm.com>
Signed-off-by: Aishwarya R <aisr@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>