UPSTREAM:wifi:CP:ath12k: Clear dangling entry in RB tree on BSS free
In the cfg80211 subsystem, all scan results of a BSS are stored in an RB tree to
facilitate easier search operations. Each BSS entry's reference count is
incremented during various association parameter updates and decremented
after usage. When the reference count reaches zero, the BSS is deleted.

However, a corner case was identified where a BSS entry could be freed before its
entry is cleared from the RB tree. This premature deletion leads to a crash due to
a dangling pointer dereference when rb_erase is called.

To address this issue, the BSS entry is now erased from the RB tree before it is
freed. This ensures that no attempts are made to access the BSS after it has been
freed, thereby preventing crashes.

Depends on : 0275-UPSTREAM-CP-wifi-cfg80211-fix-bss-rbn-double-erase-i.patch
Co-developed-by: Saravanakumar Duraisamy <quic_saradura@quicinc.com>
Change-Id: I40cbaacda24d066623a88f55ba857dd567b80135
Signed-off-by: Sowmiya Sree Elavalagan <quic_ssreeela@quicinc.com>
Signed-off-by: Narendhar Reddy <nare@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>