UPSTREAM: CP: wifi: ath12k: Enable QoS support in the Firmware
The capability to support QoS is advertised by the FW in the service
ready event and based on QoS support enablement in the host the feature
is enabled in the Firmware.

The QoS parameters that are configured for Downlink needs to be
configured in the Firmware which is used by the Firmware scheduler in
the Downlink Datapath.

The QoS parameters that are configure in UL are used to send the UL
triggers.

Add WMI support to configure the Downlink QoS parameters in the Firmware
Add WMI support to configure the Uplink QoS parameters in the Firmware

Change-Id: Icf675562273772082c8b87604ddcd679d384d4d5
Signed-off-by: Vivek Chettri <vchettri@qti.qualcomm.com>
Signed-off-by: Aishwarya R <aisr@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>