Upstream:DP:wifi: ath12k: Remove skb_list and use list for processing skb
Currently, skb list is used to store skbs reaped from the monitor
destination ring. This approach works for skbs but is not suitable for
the new frag approach, where frags are allocated instead of skbs. The
skb_list cannot be used to store frags, leading to inefficiencies and
potential issues in the monitor ring reaping path.

To address this, the existing skb list usage is removed, and a local list
is implemented to collect all monitor descriptors. This local list is then
used to process individual skbs from the monitor descriptors, ensuring
compatibility with the frag approach and improving the overall efficiency
of the monitor ring reaping process.

Patch-dependency:
0475-01-Upstream-DP-wifi-ath12k-add-monitor-desc-pool-suppor.patch

Change-Id: I970236c8bcaad2ef78a61c68bea3aca4392c7d40
Signed-off-by: P Praneesh <ppranees@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>