Upstream: DP: wifi: ath12k: Utilize monitor descriptor during processing
Currently during processing the monitor destination ring, driver caches
the skb and the length of the buffer in the monitor software descriptor.
When processing, the skb is passed as a separate member to the functions
that require it. Passing the skb as a separate member to functions is
redundant and inefficient, as the monitor descriptor already contains this
information. This approach complicates the code and makes it harder to
maintain, especially with upcoming changes related to monitor page
fragments. Modify the code to utilize the monitor software descriptor to
fetch the skb and length fields directly. This change simplifies the
function signatures and prepares the codebase for future enhancements
related to monitor page fragments.

Patch-dependency:
0475-02-Upstream-DP-wifi-ath12k-Remove-skb_list-and-use-list.patch

Change-Id: Ib1ce9e5cf43e6853a257c31e47c2e81d60773d54
Signed-off-by: P Praneesh <ppranees@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>