UPSTREAM:PROTOCOL: wifi: cfg80211: encapsulate add_links in a union for future extensibility
Encapsulate the add_links array in the cfg80211_ml_reconf_req
structure using a union to improve structural flexibility. Previously,
add_links was a fixed member, which limited the structure's
extensibility.

Update all references to add_links in cfg80211, mac80211, and nl80211
to use the new union member.

This refactoring enhances maintainability and prepares the structure
for future extensions, such as support for AP MLD in ML reconfiguration.

Change-Id: Ia3259e78b7736cd1cbd3c24c9d3463b5056f4e25
Signed-off-by: Hari Naraayana Desikan Kannan <hnaraaya@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>