QSDK:DP:wifi:ath12k:Add support to update an FSE entry
When an FSE entry has to be updated, the current support is only to
delete it and add it again with a different metadata. The reason being,
for a given tuple_info, the entry cannot be modified/added again if it
exists in the FST.

This would pose a limitation in scenarios, where FSE entry has to be
modified for the same tuple_info.

Add support to check for duplicate FSE entry based on the tuple_info
as well as rx_flow_info parameters.

If the FSE entry is same as the one requested, do not re-program; else
update the metadata with the one that is to be programmed in addition
to the existing one.

This ensures, FSE entries can be modified by multiple entities which
rely on FSE metadata for the same tuple flow.

Patch-dependency: 0132-0002-QSDK-wifi-ath12k-Add-helper-functions-for-calculatin.patch
Patch-work: NA
Tag: FSE

Change-Id: Ic35ba188320bb6a5ac3bcede6bb40bdc061c892c
Signed-off-by: Ramanathan Choodamani <rchoodam@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>