UPSTREAM:DP wifi:ath12k: Set BIT(9) in pre-reset start command based on reset context
As part of the UMAC reset sequence, the host sends a pre-reset start
command to the firmware. BIT(9) in this command is used to indicate
whether the reset is due to a UMAC hang.

This change ensures that BIT(9) is set appropriately based on the type
of reset:
- For SSR (Subsystem Restart) recovery, BIT(9) is cleared.
- For other UMAC reset scenarios (e.g., Dynamic WSI Bypass), BIT(9) is set.

The decision is made based on the UMAC reset RX action received from
the firmware.

Change-Id: I27c8273974f617795d017b84be8b059b3a706e66
Signed-off-by: Vignesh C <vignc@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>