QSDK: DP:wifi: ath12k: Add reinject handler for ppe-ds flow
Add support for reinjecting frames in the direct switch
transmit completion path.

FW exception always occurs for the initial few sdwf packets.
In the case of DS, since the FW doesn't get the svc id,
it reinjects the packet back to the host, which then
enqueues it to TCL with all necessary sdwf fields populated.

Change-Id: I40ed9b6cc986e1b16882d61464e52b3796fd0361
Signed-off-by: Prateek Sharma <prsha@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>