Upstream: DP: wifi: ath12k: Force decap type to RAW for specific frame types in monitor mode
Currently, the decap type is derived from the MPDU_START TLV and remains
fixed throughout the MPDU processing in the monitor path. However, when
the raw_mpdu bit is set in the MPDU_START TLV, the decap type must be
overridden to RAW mode.

Additionally, for management frames, control frames, and NULL data
frames, force the decap type to RAW. This ensures that monitor mode
applies the correct handling logic for these frame types, which is not
applicable for the non-RAW decapsulation.

Patch-dependency:
0475-09-QSDK-DP-wifi-ath12k-add-memory-stats-for-monitor-des.patch

Change-Id: I57aba40cd125eb43837790e4d9bb8e97deacdf8d
Signed-off-by: P Praneesh <ppranees@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>