Upstream: DP: wifi: ath12k: collect 802.11 header and payload for each MSDU
Monitor mode currently lacks support for processing packets in Ethernet
format due to the absence of 802.11 header reconstruction. To address this,
the driver is updated to handle both RX Header TLV and Buffer Address TLV
provided by the hardware.

When the RX header TLV is received, the driver allocates a SKB and
appends the status fragment containing RX Header. Upon receiving the
Buffer Address TLV, the driver appends the actual packet payload as a
fragment to the skb. A new structure, hal_rx_mon_msdu_info, is introduced
to track the first and last buffer fragments, which is essential for
correctly reassembling the complete MPDU.

These changes lay the groundwork for delivering fully reconstructed
packets to mac80211. Full re-stitching support will be introduced in
subsequent patches.

Patch-dependency:
0495-02-Upstream-DP-wifi-ath12k-Add-helper-to-safely-retriev.patch

Change-Id: I28bad72fa9a0b36753ba865d7354a0fffb81462c
Signed-off-by: P Praneesh <ppranees@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>