UPSTREAM: mac80211: Apply radio mask changes while the network interface is running
In case of  2.4 GHz + 5 GHz MLO, the 5G 40 MHz scan failed because the
radio mask was not set due to netif running check in mac80211, which
led to the VAP failing to come up.

To address this, introduced the changes to ensure the radio mask
is applied even when the interface is active, since dynamic link
addition/removal is supported in MLO scenarios.

Change-Id: I659c86529a67c1cca642350f7881c22f856e0af5
Signed-off-by: Balamurugan Ramar <bramar@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>