QSDK:wifi:ath12k: populate ar->radio_idx and use that in debugfs_register
Currently the symlinks in /sys/kernel/debug/ieee80211/ath12k_hw%d is using
ar->hw_link_id instead of the ar index in ah->radio[]. But the test scripts
will rely on the iw dev index value which is advertised based on the ar index
in ah->radio[].

To resolve this save the index position in ar->radio_idx and reused that in
debugfs_register.

Also change the parameter name from hw_link_id to radio_id to avoid ambiguity.

NOTE: This patch to be removed once the new design changes are downstreamed.

Change-Id: Ida9025aacf9aa26ab1fb4c115229327bdf685854
Signed-off-by: Avula Sri Charan <asrichar@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>