mac80211-extn: Add 240MHz minimal patch to integrate 240MHz extensions with mac80211
240MHz is a QCN feature to support higher bandwidth in a 5GHz band.
Add 240MHz feature in the qca-driver-extns/subsys/src folder. Use this
minimal patch to link the 240MHz extension code with the mac80211
component in the backports.

This minimal patch integrates the following code with the main mac80211
component:

1. Since, 240MHz is Q-Q feature, add support to advertise this capability
through vendor IE in beacon, probe and association frames for the STA.
2. Parse QCN 240 MHz vendor element and store in struct ieee802_11_elems.
This data will be used for Q-Q 240 MHz association in 5 GHz.
3. Get the associated STA 240MHz capabilities from NL Vendor Data
and then override the STA phy capabilities with copied one during STA
association at the AP side.
4. Store Q-Q 240 MHz vendor IE in the bss data structure for STA mode
and add it in the Association Request frame sent to AP.
5. Use same data to upgrade from 160 MHz to 320 MHz (with punturing
bitmap) while generating channel data for the link.
6. Peer association parameters sent to firmware via WMI will reflect
these bandwidth changes as well.
7. EHT connections on the 5GHz band currently do not set a bandwidth limit,
which may prevent the use of 240MHz operation when puncturing is enabled.
The default behavior does not explicitly allow 240MHz, as the bw_limit
remains unset for 5GHz EHT.
8. Set the bw_limit to IEEE80211_CONN_BW_LIMIT_320 for 5GHz EHT connections
to enable 240MHz operation. This ensures that the connection logic
correctly reflects the capabilities of EHT in the 5GHz band.

Change-Id: Ie74b2e843521b2a9ecabd195c36a53ae278983e0
Signed-off-by: Venkateswara Swamy Bandaru <vbandaru@qti.qualcomm.com>
Signed-off-by: Sri Vidya Gunturi <sgunturi@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>