ath12k-extn: Add support to compile ath12k extension code for ath12k
Add support to compile the ath12k extension files using the
ath12k/Makefile. The qcn_extns/Makefile is linked to ath12k/Makefile and
the qcn_extns/Kconfig is linked to the ath12k/Kconfig.

Add this patch in the qca-driver-extns/patches folder. Copy and apply
this patch into the backports for building the extension code.

Change-Id: I1edbaa8aa3f5d5a40cf71f4f0fc66436c43e6221
Signed-off-by: Sri Vidya Gunturi <sgunturi@qti.qualcomm.com>
Signed-off-by: Venkateswara Swamy Bandaru <vbandaru@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>