Upstream:DP: Add logic for PPDU based reaping for monitor destination ring
Currently, the monitor destination ring is reaped by adding individual
monitor descriptors to a local list called `ppdu_desc_used_list`. However,
a single PPDU (Physical Protocol Data Unit) can span across multiple NAPI
poll cycles. The existing logic does not correctly accumulate all the
descriptors belonging to a single PPDU when it is split across multiple
budgets. This leads to undefined behavior during monitor descriptor
processing, as incomplete PPDUs may be processed prematurely.

To resolve this, introduce a PPDU-based reaping mechanism for the monitor
destination ring. With this approach, the driver accumulates all status
buffers associated with a PPDU and appends them to the ppdu_desc_used_list
only when the end-of-PPDU flag is detected in the monitor destination ring
descriptor. Once a complete PPDU is collected, the used list is transferred
to the `ppdu_desc_proc_list` for TLV parsing and further processing. After
processing, the descriptors are returned to the `ppdu_desc_free_list`.

This change ensures correctness in PPDU handling and prepares the driver
for future enhancements in monitor processing.

Patch-dependency:
0475-07-UPSTREAM-DP-wifi-ath12k-Enable-low-threshold-interru.patch

Change-Id: I57f727eb5ae05f090917f4b5232b325ca18e40de
Signed-off-by: P Praneesh <ppranees@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>