UPSTREAM:CP: wifi: ath12k: Enable pdev suspend and resume command
The pdev suspend command helps reduce power consumption. When the last
vdev on a particular pdev is deleted, the host sends a
WMI_PDEV_SUSPEND_CMDID to the firmware. In response, the firmware sends
back a WMI_PDEV_SUSPEND_EVENTID to the host.

Similarly, when the first vdev is created and the corresponding pdev is
in suspend mode, the host sends a WMI_PDEV_RESUME_CMDID to the firmware.
The firmware then responds with a WMI_PDEV_RESUME_EVENTID.

Patch-dependency: "UPSTREAM:CP:wifi:ath12k: ErP support"
Patch-work: None

Change-Id: I41a300804110669503eca800ea88687862e6948f
Signed-off-by: Aaradhana Sahu <aaradhana.sahu@oss.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>