QSDK:CP:wifi:ath12k: Ensure arvif is created before updating splitphy bank ID
Currently, during split PHY configuration, the driver iterates over all
links in the `links_map` and updates the `splitphy_ds_bank_id` for each
associated `iter_arvif`. However, this logic does not check whether the
link has been fully created (`is_created`), which can lead to unintended
access to uninitialized or invalid link contexts.

Fix this by adding a check for `!iter_arvif->is_created` in the loop
to skip the link if it is not created yet. Once the chanctx is assigned
for that link, we will eventually iterate over that link again

Patch-dependency:
	0337-01-QSDK-DP-wifi-Add-support-for-PPE-DS-Direct-switch-fe.patch

Change-Id: I54adec09cfab8b9e37f2961d7f41ff679e394764
Signed-off-by: Vignesh C <vignc@qti.qualcomm.com>
Signed-off-by: Ganesh Kariganuru Mahabalesh <gkarigan@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>