QSDK: Platform: wifi:ath12k: Add support for MLO reconfig QMI message
This patch introduces a new QMI message, QMI_MLO_RECONFIG (0x005F),
to update the adjacency information of targets to the firmware when
a target is bypassed or brought back from the bypassed state.

It also adds the corresponding request and response handlers for
this message.

To support both the maximum capabilities in the Host Capability QMI
message and the bypass configuration in the MLO Reconfig QMI message
during SSR scenarios, we need to maintain two instances of WSI
information.

Introduce bypass_wsi_info to store and use the WSI data specifically
for bypass cases.

Patch-dependency: 0325-02-wifi-ath12k-Add-support-for-four-link-MLO-in-AP-MLD.patch
Change-Id: I3eda141eb00fe312dc110a170425c0ae065ca90a
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>