wifi:ath12k: Enable static WSI bypass with module param
The static WSI bypass mechanism will bypass one radio only if two active radios
are present in the Access Point (AP). This is controlled using a module
parameter and is configured during the insertion (insmod) of the ath12k module.

The module parameter uses the WSI index of each radio as a bitmap to
determine which radio to bypass. Below are examples of how
to use this parameter:

Examples:

wsi_bypass_bmap = 1  // 0x0001
This will bypass the 2G radio interface, as the WSI index of the 2G radio is 0.

wsi_bypass_bmap = 2  // 0x0010
This will bypass the 5G radio interface, as the WSI index of the 5G radio is 1.

The current implementation leverages the dynamic bypass infrastructure to disable
a radio when no active VAPs (Virtual Access Points) are present
on the bypassed radio. The static bypass logic checks whether the static
bypass bitmap is set for a given radio and initiates dynamic bypass
for that radio once all chips are active.

Note: After a radio has been bypassed, re-adding it is only supported by reloading
the module (i.e., performing rmmod followed by insmod).

Change-Id: I13fbee21aafcee2c1ae7817027e1aa4b43312fb7
Signed-off-by: Gautham Kumar Senthilkumaran <gauthamk@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>