UPSTREAM:CP|DP:wifi: ath12k: Skip stopping hw_queues in case of Mode-0 recovery.
As per upstream design In case of Mode-0 recovery the hw_quques
are stopped after that an mgmt frame coming in tx path is being
added into per CPU pending queue.

Now at the end of recovery when ath12k does hw_wake_queues() it
will check the pending queue and queue that frame immediatley
because of which after recovery disassoc frame is going to FW
before STA associates which leads to a FW crash.

Ideally when Mode-0 recovery is going on all mgm tx frames
should be freed as part of drv_tx instead of retaining in
pending queue and queue it later after Recovery.

To achieve that skip the ieee80211_stop_queues() in case of
Mode-0 recovery and queue all frames to drv_tx().

Now inside ath12k_wifi7_mac_op_tx() when recovery is in
progress these frames will be dropped.

Change-Id: I5772b8a164619b4e1552f48a066cbdaca814ea5b
Signed-off-by: Avula Sri Charan <asrichar@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>