UPSTREAM: DP: wifi: ath12k: Fix UMAC migration issue in DP
In the current implementation of primary link migration, TID setup is
performed for the new primary peer, and TID deletion is done for the
old primary peer.

However, in the Unified Driver architecture, rx_tid is maintained under
dp_peer. This causes the existing approach to result in memory
corruption due to improper handling of shared structures.

To address this, the new implementation avoids reinitializing TIDs
during migration. Instead, it updates the qref table with the new SOC
and clears the qref table associated with the old SOC, ensuring safe and
consistent state transitions.

Fixes: I319730cf68d5 ("wifi: ath12k: add support for primary link
migration")

Change-Id: Icaf6856bfd742c26b782e6d54a28aa1a72520ebc
Signed-off-by: Girish Kumar M P <gmp@qti.qualcomm.com>
Signed-off-by: Balamurugan S <bselvara@qti.qualcomm.com>